Tests/LibWeb/Text/expected/cookie.txt

   Basic test: "cookie=value"
Multiple cookies: "cookie1=value1; cookie2=value2; cookie3=value3"
Nameless cookie: "value"
Valueless cookie: "cookie="
Nameless and valueless cookie: ""
Invalid control character: ""
Non-ASCII domain: ""
Secure cookie prefix: ""
Host cookie prefix: ""
Large value: "cookie=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
Overly large value: ""
HTTP only: ""
Public suffix: ""
SameSite=Lax: "cookie=value"
SameSite=Strict: "cookie=value"
SameSite=None: ""
Max-Age (before expiration): "cookie-max-age=value"
Expires (before expiration): "cookie-expires=value; cookie-max-age=value"
Max-Age (after expiration): ""
Expires (after expiration): ""
Max-Age in past: ""
Expires in past: ""
Invalid expiry (date does not exist): "cookie=value"
Invalid expiry (missing time): "cookie=value"
Invalid expiry (missing day): "cookie=value"
Invalid expiry (missing month): "cookie=value"
Invalid expiry (missing year): "cookie=value"
